# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Machinedog Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/clients:
    get:
      operationId: listClients
      tags: [admin]
      summary: List all clients (admin only)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: All clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientList"
  /admin/clients/invite:
    post:
      operationId: inviteClient
      tags: [admin]
      summary: Invite a new client by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteClientBody"
      responses:
        "200":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
  /admin/clients/{id}:
    get:
      operationId: getClientById
      tags: [admin]
      summary: Get a client by id (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [admin]
      summary: Delete a client (admin only). Hard-deletes invited rows with no history; soft-suspends clients with related records.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteClientResponse"
  /admin/clients/{id}/resend-invite:
    post:
      operationId: resendClientInvite
      tags: [admin]
      summary: Regenerate the invite token for a client and resend the invite email
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
  /admin/clients/{id}/balance:
    patch:
      operationId: adjustClientBalance
      tags: [admin]
      summary: Manually adjust a client token balance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
...
  /admin/projects/{id}/owner:
    patch:
      operationId: reassignProjectOwner
      tags: [admin]
      summary: Reassign a project to a different client (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignProjectOwnerBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /admin/projects/import-github:
    post:
      operationId: importGithubProject
      tags: [admin]
      summary: Create a Machinedog project from a GitHub repo (admin only)
      description: |
        Creates a new project owned by the calling admin's client record, with
        githubOwner / githubRepo / githubDefaultBranch pre-populated. Admin can
        then reassign owner or attach collaborators via the existing controls.
        Returns 409 if a project with that owner/repo already exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportGithubProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          description: A project for this repo already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Project secrets ───────────────────────────────────────────────────────
...
  /admin/orders:
    get:
      operationId: listAllBuildOrders
      tags: [admin]
      summary: List paid build deposits and retainer subscriptions (admin only)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: All build orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildOrderList"

# ─── Schemas ───────────────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        tokenBalance:
          type: ["integer", "null"]
      required: [error]
    Client:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        email:
          type: string
        tokenBalance:
          type: integer
        totalTokensUsed:
          type: integer
        isAdmin:
          type: boolean
        stripeCustomerId:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, invited]
        portalSubscriptionStatus:
          type: string
          enum: [none, trialing, active, past_due, canceled, incomplete]
        portalCurrentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        # The organization the client belongs to. Optional because legacy
        # clients may not yet have a backfilled organization row. The
        # frontend uses `organization.planType` to gate plan-restricted
        # surfaces (e.g. healthcare templates).
        organization:
          type: ["object", "null"]
          properties:
            id:
              type: string
            name:
              type: string
            planType:
              type: string
              enum: [free, starter, pro, healthcare, enterprise]
          required: [id, name, planType]
      required: [id, email, tokenBalance, totalTokensUsed, isAdmin, status, portalSubscriptionStatus, createdAt]
    ClientList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
      required: [data, total]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: string
        summary:
          type: string
        liveUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, completed, archived]
        projectType:
          type: string
          enum: [web_app, mobile_app, api, internal_tool, healthcare_template, other]
        framework:
          type: string
        templateSlug:
          type: ["string", "null"]
        healthcareMode:
          type: boolean
        phiAllowed:
          type: boolean
        baaStatus:
          type: string
          enum: [not_required, required, pending, active, expired]
        workspaceProvider:
          type: string
          enum: [database, replit, external]
        workspaceStatus:
          type: string
          enum: [idle, starting, running, stopping, error]
        workspaceUrl:
          type: ["string", "null"]
        lastStartedAt:
          type: ["string", "null"]
          format: date-time
        lastStoppedAt:
          type: ["string", "null"]
          format: date-time
        consultingBookingId:
          type: ["integer", "null"]
        githubOwner:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
        githubDefaultBranch:
          type: string
        previewUrlTemplate:
          type: ["string", "null"]
        productionUrl:
          type: ["string", "null"]
        operatorEmail:
          type: ["string", "null"]
        # Per-project token used by the heartbeat snippet installed in the client
        # app to auto-report its current Replit dev URL. Returned to owners only.
        heartbeatToken:
          type: ["string", "null"]
        # ISO timestamp of the last heartbeat received. null = snippet has never
        # checked in (UI shows "not connected").
        heartbeatAt:
          type: ["string", "null"]
          format: date-time
        viewerRole:
          type: string
          enum: [owner, collaborator]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clientId, title, description, summary, status, projectType, framework, healthcareMode, phiAllowed, baaStatus, workspaceProvider, workspaceStatus, githubDefaultBranch, viewerRole, createdAt, updatedAt]
...
    ReassignProjectOwnerBody:
      type: object
      properties:
        clientId:
          type: integer
      required: [clientId]
...
    ImportGithubProjectBody:
      type: object
      properties:
        owner:
          type: string
          minLength: 1
        repo:
          type: string
          minLength: 1
        defaultBranch:
          type: string
          minLength: 1
        title:
          type: string
          description: |
            Optional override. Defaults to the repo name with hyphens replaced
            by spaces and Title Cased.
      required: [owner, repo, defaultBranch]
...
    BuildOrder:
      type: object
      properties:
        id:
          type: integer
        kind:
          type: string
          enum: [build, retainer]
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        amountCents:
          type: integer
        currency:
          type: string
        source:
          type: string
        status:
          type: string
          enum: [pending, completed, active, cancelled]
        stripeSessionId:
          type: string
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        stripePaymentIntentId:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - kind
        - amountCents
        - currency
        - source
        - status
        - stripeSessionId
        - createdAt
        - updatedAt
    BuildOrderList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BuildOrder"
        total:
          type: integer
      required: [data, total]
...
    InviteClientBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    InviteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    DeleteClientResponse:
      type: object
      properties:
        success:
          type: boolean
        mode:
          type: string
          enum: [hard, soft]
          description: hard = row removed; soft = client suspended due to existing history
        message:
          type: string
      required: [success, mode, message]
    AdjustBalanceBody:
      type: object
      properties:
        delta:
          type: integer
          description: Positive to add tokens, negative to deduct
        reason:
          type: string
      required: [delta]
...